basic program

+++ personal_details.cpp
@@ -4,7 +4,7 @@ int main(){
     int reg_no = 12221533;
     char name[10] = "Atul";
     float marks = 0.01;
-    double phone_no = 123456789;
+    long long int phone_no = 123456789;
     cout << reg_no << endl;
     cout << name << endl;
     cout << marks << endl;
